Migrate MultibindingGraphWithWithScopedMapProviderDeps test
